+++ src/features/business/presentation/businessdetails/BusinessDetailsUiModel.ts
@@ -1,8 +1,3 @@
-import { Business } from '../../domain/model/Business';
-import { Review } from '../../domain/model/Review';
-import { User } from '../../domain/model/User';
-import * as BusinessHelper from '../helper/BusinessHelper';
-
 export interface BusinessDetailsUiModel {
   id: string;
   name: string;
@@ -27,36 +22,3 @@ export interface UserUiModel {
   name: string;
   imageUrl: string;
 }
-
-export const toUiModel = (business: Business): BusinessDetailsUiModel => {
-  return {
-    id: business.id,
-    name: business.name.toUpperCase(),
-    imageUrl: business.imageUrl,
-    ratingImage: BusinessHelper.getRatingImage(business.rating),
-    reviewCount: `${business.reviewCount} reviews`,
-    address: business.address,
-    priceAndCategories: BusinessHelper.formatPriceAndCategories(business.price, business.categories),
-    //
-    reviews: toReviewUiModels(business.reviews!),
-  };
-};
-
-const toReviewUiModels = (reviews: Review[]): ReviewUiModel[] => {
-  return reviews.map(review => {
-    return {
-      id: review.id,
-      user: toUserUiModel(review.user),
-      text: review.text,
-      ratingImage: BusinessHelper.getRatingImage(review.rating),
-      timeCreated: review.timeCreated,
-    };
-  });
-};
-
-const toUserUiModel = (user: User): UserUiModel => {
-  return {
-    name: user.name,
-    imageUrl: user.imageUrl,
-  };
-};
